refactor sharding components
