+++ src/MaxText/train.py
@@ -50,6 +50,7 @@ from MaxText import maxtext_utils
 from MaxText import train_utils
 from MaxText import profiler
 from MaxText import pyconfig
+from MaxText import sharding
 from MaxText.layers.multi_token_prediction import calculate_mtp_acceptance_rate, calculate_mtp_loss
 from MaxText.common_types import ShardMode
 from MaxText.data_loader import DataLoader
@@ -141,7 +142,7 @@ def loss_fn(model, config, data, dropout_rng, params, is_train=True):
     else:
       one_hot_targets = jax.nn.one_hot(data["targets"], config.vocab_size)
       xent, _ = max_utils.cross_entropy_with_logits(logits, one_hot_targets)
-      xent = maxtext_utils.maybe_shard_with_logical(
+      xent = sharding.maybe_shard_with_logical(
           xent,
           ("activation_embed_and_logits_batch", "activation_length"),
           model.mesh,
@@ -374,9 +375,7 @@ def train_loop(config, recorder, state=None):
       state = _merge_dpo_state(state, reference_params)
     state_mesh_shardings = _merge_dpo_state(state_mesh_shardings, state_mesh_shardings.params["params"])
 
-  params_shardings, state_mesh_shardings = maxtext_utils.maybe_update_params_sharding_with_opt(
-      config, state_mesh_shardings
-  )
+  params_shardings, state_mesh_shardings = sharding.maybe_update_params_sharding_with_opt(config, state_mesh_shardings)
 
   p_train_step, p_eval_step = train_utils.jit_train_and_eval_step(
       config, model, mesh, state, state_mesh_shardings, train_step, eval_step, eval_data_iterator, params_shardings
@@ -385,7 +384,7 @@ def train_loop(config, recorder, state=None):
   with mesh, nn_partitioning.axis_rules(config.logical_axis_rules):
     shaped_batch = maxtext_utils.get_shaped_batch(config)
     if config.shard_optimizer_over_data:
-      state = maxtext_utils.maybe_shard_with_name(state, state_mesh_shardings, config.shard_mode)
+      state = sharding.maybe_shard_with_name(state, state_mesh_shardings, config.shard_mode)
     compiled = p_train_step.lower(state, shaped_batch, init_rng).compile()
     compiled_stats = compiled.memory_analysis()
     max_utils.print_compiled_memory_stats(compiled_stats)
@@ -410,7 +409,7 @@ def train_loop(config, recorder, state=None):
         with maybe_record_goodput(recorder, GoodputEvent.STEP, step):
           with mesh, nn_partitioning.axis_rules(config.logical_axis_rules):
             if config.shard_optimizer_over_data:
-              state = maxtext_utils.maybe_shard_with_name(state, state_mesh_shardings, config.shard_mode)
+              state = sharding.maybe_shard_with_name(state, state_mesh_shardings, config.shard_mode)
             state, metrics = p_train_step(state, example_batch, nextrng)
 
       step_time_delta = datetime.datetime.now() - last_step_completion
